+++ Packages/ai.videokit.videokit/Runtime/Sources/AudioComponentSource.cs
@@ -20,48 +20,26 @@ namespace VideoKit.Sources {
         /// <summary>
         /// Create an audio buffer source from an AudioListener.
         /// </summary>
-        /// <param name="handler">Handler to receive audio buffers.</param>
-        /// <param name="listener">Audio listener to record.</param>
-        public AudioComponentSource (Action<AudioBuffer> handler, AudioListener listener) : this(
-            handler,
-            default,
-            listener.gameObject
-        ) { }
-
-        /// <summary>
-        /// Create an audio buffer source from an AudioSource.
-        /// </summary>
-        /// <param name="handler">Handler to receive audio buffers.</param>
-        /// <param name="source">Audio source to record.</param>
-        public AudioComponentSource (Action<AudioBuffer> handler, AudioSource source) : this(
-            handler,
-            default,
-            source.gameObject
-        ) { }
-
-        /// <summary>
-        /// Create an audio buffer source from an AudioListener.
-        /// </summary>
-        /// <param name="handler">Handler to receive audio buffers.</param>
-        /// <param name="clock">Clock for generating timestamps. Can be `null` if recorder does not require timestamps.</param>
         /// <param name="listener">Audio listener for the current scene.</param>
-        public AudioComponentSource (Action<AudioBuffer> handler, IClock? clock, AudioListener listener) : this(
-            handler,
-            clock,
-            listener.gameObject
-        ) { }
+        /// <param name="handler">Handler to receive audio buffers.</param>
+        /// <param name="clock">Clock for generating timestamps.</param>
+        public AudioComponentSource (
+            AudioListener listener,
+            Action<AudioBuffer> handler,
+            IClock? clock = null
+        ) : this(listener.gameObject, handler, clock) { }
 
         /// <summary>
         /// Create an audio buffer source from an AudioSource.
         /// </summary>
-        /// <param name="handler">Handler to receive audio buffers.</param>
-        /// <param name="clock">Clock for generating timestamps. Can be `null` if recorder does not require timestamps.</param>
         /// <param name="source">Audio source to record.</param>
-        public AudioComponentSource (Action<AudioBuffer> handler, IClock? clock, AudioSource source) : this(
-            handler,
-            clock,
-            source.gameObject
-        ) { }
+        /// <param name="handler">Handler to receive audio buffers.</param>
+        /// <param name="clock">Clock for generating timestamps.</param>
+        public AudioComponentSource (
+            AudioSource source,
+            Action<AudioBuffer> handler,
+            IClock? clock = null
+        ) : this(source.gameObject, handler, clock) { }
 
         /// <summary>
         /// Close the media source and release resources.
@@ -73,7 +51,7 @@ namespace VideoKit.Sources {
         #region --Operations--
         private readonly AudioSourceAttachment attachment;
 
-        private AudioComponentSource (Action<AudioBuffer> handler, IClock? clock, GameObject gameObject) {
+        private AudioComponentSource (GameObject gameObject, Action<AudioBuffer> handler, IClock? clock) {
             var sampleRate = AudioSettings.outputSampleRate;
             attachment = gameObject.AddComponent<AudioSourceAttachment>();
             attachment.sampleBufferDelegate = (data, channels) => {